migrate palette to @guardian/source-foundations
