+++ packages/@guardian/source-foundations/src/index.ts
@@ -4,3 +4,32 @@ export { transitions } from './animation';
 // breakpoints
 export type { Breakpoint } from './breakpoints';
 export { breakpoints } from './breakpoints';
+
+// palette
+export {
+	brand,
+	brandAlt,
+	neutral,
+	error,
+	success,
+	news,
+	opinion,
+	sport,
+	culture,
+	lifestyle,
+	labs,
+	specialReport,
+	focus,
+	background,
+	brandBackground,
+	brandAltBackground,
+	border,
+	brandBorder,
+	brandAltBorder,
+	line,
+	brandLine,
+	brandAltLine,
+	text,
+	brandText,
+	brandAltText,
+} from './palette';
