+++ src/core/foundations/src/palette.ts
@@ -1,0 +1,1 @@
+../../../../packages/@guardian/source-foundations/src/palette.ts
