--- src/core/foundations/src/palette/background/brand-alt.ts
@@ -1,10 +1,0 @@
-import { neutral, brandAlt } from '../global';
-
-export const brandAltBackground = {
-	primary: brandAlt[400],
-	ctaPrimary: neutral[7],
-	ctaPrimaryHover: '#454545',
-	ctaSecondary: brandAlt[200],
-	ctaSecondaryHover: '#F2AE00',
-	ctaTertiaryHover: '#FFD213',
-};
