--- src/core/foundations/src/palette/background/index.ts
@@ -1,3 +1,0 @@
-export { background } from './default';
-export { brandBackground } from './brand';
-export { brandAltBackground } from './brand-alt';
