--- src/core/foundations/src/palette/border/brand-alt.ts
@@ -1,5 +1,0 @@
-import { neutral } from '../global';
-
-export const brandAltBorder = {
-	ctaTertiary: neutral[7],
-};
