--- src/core/foundations/src/palette/border/index.ts
@@ -1,3 +1,0 @@
-export { border } from './default';
-export { brandBorder } from './brand';
-export { brandAltBorder } from './brand-alt';
