--- src/core/foundations/src/palette/index.ts
@@ -1,5 +1,0 @@
-export * from './global';
-export { background, brandBackground, brandAltBackground } from './background';
-export { border, brandBorder, brandAltBorder } from './border';
-export { line, brandLine, brandAltLine } from './line';
-export { text, brandText, brandAltText } from './text';
