--- src/core/foundations/src/palette/line/brand-alt.ts
@@ -1,5 +1,0 @@
-import { neutral } from '../global';
-
-export const brandAltLine = {
-	primary: neutral[7],
-};
