--- src/core/foundations/src/palette/line/brand.ts
@@ -1,5 +1,0 @@
-import { brand } from '../global';
-
-export const brandLine = {
-	primary: brand[600],
-};
