--- src/core/foundations/src/palette/line/default.ts
@@ -1,5 +1,0 @@
-import { neutral } from '../global';
-
-export const line = {
-	primary: neutral[86],
-};
