--- src/core/foundations/src/palette/line/index.ts
@@ -1,3 +1,0 @@
-export { line } from './default';
-export { brandLine } from './brand';
-export { brandAltLine } from './brand-alt';
