--- src/core/foundations/src/palette/text/brand-alt.ts
@@ -1,10 +1,0 @@
-import { neutral } from '../global';
-
-export const brandAltText = {
-	primary: neutral[7],
-	supporting: neutral[60],
-	ctaPrimary: neutral[100],
-	ctaSecondary: neutral[7],
-	ctaTertiary: neutral[7],
-	anchorPrimary: neutral[7],
-};
