--- src/core/foundations/src/palette/text/index.ts
@@ -1,3 +1,0 @@
-export { text } from './default';
-export { brandText } from './brand';
-export { brandAltText } from './brand-alt';
